connect user's id to todo, modify user & todo models and routes

+++ routes/todos.js
@@ -1,11 +1,26 @@
 const express = require("express");
 const router = express.Router();
+const jwt = require("jsonwebtoken");
 const { Todo } = require("../models/todos");
+
+// handle authentication --> verify JWT
+router.use((req, res, next) => {
+  const authHeader = req.headers["authorization"];
+  const token = authHeader ? authHeader.split(" ")[1] : null;
+
+  if (token == null) return res.sendStatus(401);
+
+  jwt.verify(token, process.env.SECRET, (error, decodedToken) => {
+    if (error) return res.sendStatus(403);
+    req.user = decodedToken.userId;
+    next();
+  });
+});
 
 // Get all todos data from database
 router.get("/", async (req, res) => {
   try {
-    const todos = await Todo.find();
+    const todos = await Todo.find({ userId: req.user });
     res.json({ todos });
   } catch (error) {
     res.status(500).json({ message: "Failed to get all todos" });
@@ -16,7 +31,8 @@ router.get("/", async (req, res) => {
 router.post("/new", async (req, res) => {
   try {
     const { title, description, completed } = req.body;
-    const todo = new Todo({ title, description, completed });
+    const userId = req.user;
+    const todo = new Todo({ title, description, completed, userId });
     const saveTodo = await todo.save();
     res.json({ message: "Successfully created a new todo!", todo: saveTodo });
   } catch (error) {
@@ -28,7 +44,9 @@ router.post("/new", async (req, res) => {
 router.get("/:id", async (req, res) => {
   try {
     const todo = await Todo.findById(req.params.id);
-    res.json({ message: "Successfully viewed todo details!", todo });
+
+    todo.userId != req.user ? res.status(403).json({ message: "Unauthorized" })
+    : res.json({ message: "Successfully viewed todo details!", todo });
   } catch (error) {
     res.status(404).json({ message: "Failed to view todo details!" });
   }
@@ -37,13 +55,12 @@ router.get("/:id", async (req, res) => {
 // Update a todo by id
 router.patch("/:id", async (req, res) => {
   try {
-    const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, req.body, {
+    const updateTodo = await Todo.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
     });
-    res.json({
-      message: "Successfully changed the todo data!",
-      todo: updatedTodo,
-    });
+
+    updateTodo.userId != req.user ? res.status(403).json({ message: "Unauthorized" })
+    : res.json({ message: "Successfully changed the todo data!", todo: updateTodo });
   } catch (error) {
     res.status(400).json({ message: "Failed to change todo data!" });
   }
@@ -53,7 +70,10 @@ router.patch("/:id", async (req, res) => {
 router.delete("/:id", async (req, res) => {
   try {
     const todo = await Todo.findByIdAndDelete(req.params.id);
-    res.json({ message: "Successfully deleted todo!", todo });
+
+    !todo ? res.status(404).json({ message: "Todo not found" })
+    : todo.userId.toString() !== req.user.toString() ? res.status(403).json({ message: "Unauthorized" })
+    : res.json({ message: "Successfully deleted todo!", todo });
   } catch (error) {
     res.status(400).json({ message: "Failed to delete todo!" });
   }
@@ -62,7 +82,7 @@ router.delete("/:id", async (req, res) => {
 // Delete all todos
 router.delete("/", async (req, res) => {
   try {
-    await Todo.deleteMany({});
+    await Todo.deleteMany({ userId: req.user });
     res.json({ message: "Successfully deleted all todos" });
   } catch (error) {
     res.status(500).json({ message: "Failed to delete all todos" });
